Move fields to a separate file
